+++ lexicmap/cmd/index.go
@@ -107,13 +107,14 @@ Important parameters:
                             ► This value is used to remove masks with a prefix of low-complexity.
 
   --- Seeds data (k-mer-value data) ---
- *1. --seed-max-desert      ► Maximum length of distances between seeds (default: 450).
-                            The default value of 450 guarantees queries >450 bp would match at least one seed.
+ *1. --seed-max-desert      ► Maximum length of distances between seeds (default: 200).
+                            The default value of 200 guarantees queries >=200 bp would match at least one seed.
                             ► Large regions with no seeds are called sketching deserts. Deserts with seed distance
                             larger than this value will be filled by choosing k-mers roughly every
-                            --seed-in-desert-dist (150 by default) bases.
-                            ■ Smaller values improve the search sensitivity, slow down the indexing speed, increase
-                            the indexing memory occupation and increase the index size.
+                            --seed-in-desert-dist (50 by default) bases.
+                            ■ Big values decrease the search sensitivity for distant targets, speed up the indexing
+                            speed, decrease the indexing memory occupation and decrease the index size. While the
+                            alignment speed is almost not affected.
   2. -c/--chunks,           ► Number of seed file chunks (maximum: 128, default: #CPUs).
                             ► Bigger values accelerate the search speed at the cost of a high disk reading load.
                             The maximum number should not exceed the maximum number of open files set by the
@@ -319,7 +320,7 @@ Important parameters:
 
 		outputDir := outDir != ""
 		if outputDir {
-			makeOutDir(outDir, force, "out-dir")
+			makeOutDir(outDir, force, "out-dir", opt.Verbose || opt.Log2File)
 		}
 
 		// ---------------------------------------------------------------
@@ -469,9 +470,9 @@ func init() {
 
 	indexCmd.Flags().IntP("seed-min-prefix", "p", 15,
 		formatFlagUsage(`Minimum length of shared substrings (anchors) in searching. Here, this value is used to remove low-complexity masks and choose k-mers to fill sketching deserts.`))
-	indexCmd.Flags().IntP("seed-max-desert", "D", 450,
-		formatFlagUsage(`Maximum length of sketching deserts. Deserts with seed distance larger than this value will be filled by choosing k-mers roughly every --seed-in-desert-dist bases.`))
-	indexCmd.Flags().IntP("seed-in-desert-dist", "d", 150,
+	indexCmd.Flags().IntP("seed-max-desert", "D", 200,
+		formatFlagUsage(`Maximum length of sketching deserts, or maximum seed distance. Deserts with seed distance larger than this value will be filled by choosing k-mers roughly every --seed-in-desert-dist bases.`))
+	indexCmd.Flags().IntP("seed-in-desert-dist", "d", 50,
 		formatFlagUsage(`Distance of k-mers to fill deserts.`))
 
 	// ------  generate mask from the top N biggest genomes